A naive approach to solve problem of false positives on alive?
This approach tries to resolves the issue where multiple containers are
running sidekiq in a given cluster but only a single container is
actually processing work. The intent here is that you configure each
container to to use a unique liveness key and only update the alive key
when a work process a job for it's own container. This will now always
work and an approach that uses a unique key per container may be better.